pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = ["Лизанька <fedko.liza@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
faker = "^37.1.0"


[tool.poetry.group.test.dependencies]
pytest-django = "^4.11.1"
pytest-mock = "^3.14.1"
bs4 = "^0.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
